LayoutTests/platform/ios-sim-deprecated/fast/events/platform-wheelevent-in-scrolling-div-expected.txt

FAIL div.scrollTop should be 200. Was 0.
FAIL div.scrollLeft should be 120. Was 0.
